requirements.txt
gunicorn==20.1.0
Flask==2.2.2
tensorflow-cpu==2.10.0
numpy==1.23.3
pybase64==1.2.3
opencv-python-headless
Pillow